# openapi.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
paths:
  /api/game/video:
    get:
      summary: Video Feed
      operationId: video_feed_api_game_video_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/android/video:
    get:
      summary: Android Video Feed
      operationId: android_video_feed_api_android_video_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/android/image:
    get:
      summary: Android Image
      operationId: android_image_api_android_image_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResult'
  /api/android/click:
    post:
      summary: Android Click
      operationId: android_click_api_android_click_post
      parameters:
      - required: true
        schema:
          title: X
          type: integer
        name: x
        in: query
      - required: true
        schema:
          title: Y
          type: integer
        name: y
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/android/scripts:
    get:
      summary: Android Scripts
      operationId: android_scripts_api_android_scripts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Android Scripts Api Android Scripts Get
                type: array
                items:
                  type: string
  /api/android/scripts/{script}:
    post:
      summary: Android Run Script
      operationId: android_run_script_api_android_scripts__script__post
      parameters:
      - required: true
        schema:
          title: Script
          type: string
        name: script
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ImageResult:
      title: ImageResult
      required:
      - base64
      - uuid
      type: object
      properties:
        base64:
          title: Base64
          type: string
        uuid:
          title: Uuid
          type: string
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string